Added support for pivoting tools in vtkTrackerWidget and fixed bugs.
vtkTrackerWidget now has an interface to pivot tools but I also fixed some
bugs with the vtkTrackerTool such that the pivoting algorithm would add
invalid transforms to the Amoeba Optimizer.  If the tool was missing the
identity transform would be added and there were no checks as to whether
the transform was recently updated.  I changed to accept a pivoting
boolean switch turns the pivot collection on and off and instead of using
InsertNextCalibrationPoint.