BUG: Fixed the enabling and disabling of the start and stop tracking buttons.
Fixed how the start and stop tracking buttons are turned on an off and
also fixed a bug around the NDI drop down combo box for the volume
selections.  The start tracking button now turns off when configure
tracker button is pressed and only turns back on once the configuration
has ended.